Added code for telling grub to ignore a kernel update
Previously, there was a problem with kernel updates:

If the RAM Session was updated but the Original OS was NOT (through
'sudo rupdate'), the Original OS would fail to boot because it didn't
have the /lib/modules necessary to work with the new kernel.

This commit makes rupdate set an ENV variable (IGNORE_KERNEL) in cases
where the RAM Session received a kernel update but the Original OS
did not. Grub can read this variable, and not create entries for new
kernels that the Original OS can't boot anyway.

Note: This will NOT work yet, as I also need a sed line that changes
/etc/grub.d/10_linux to actually care about the message