Added quiet version of Ctrl+C trap at the beginning of script
Some files (like log files) get written to different places around the
filesystem before we start copying the entire filesystem to $DEST. If the
user exits at that point, he probably thinks that since copying didn't
start, there's nothing to clean up, so if we ask him if we should clean
up, he may think it's a bug and say no. To get around this, a quiet
version of the Ctrl+C trap will be enabled at the beginning of the
script so that if the user exits at any point before the filesystem
starts copying, we still clean up - we just don't ask the user. If we
are at that point, there was no previous RAM Session installed, and the
current RAM Session didn't even get copied to $DEST yet, so it should
be pretty safe to cleanup without asking the user.